Change default shell to bash
Per @mcovarr:

The job shell is now configurable and /bin/sh is a valid choice. We've
left the default job shell at /bin/bash to not break legacy workflows.
You can set the job shell on a systemwide basis with
system.job-shell=/bin/sh or on a per-backend basis by setting
job-shell=/bin/sh in the config for the AWS backend. The
StandardAsyncExecutionActor value of jobShell has already been populated
with the shell you should use.

But for the reason of not breaking legacy workflows cited above I don't
know that I'd recommend setting the AWS Batch backend to use /bin/sh
unless you really believe strongly it should use that. Users may find
that some of their workflows that run on other backends don't run on AWS
Batch due to the shell discrepancy. They could update their configs to
use /bin/bash of course, but having to figure that out probably wouldn't
be the best experience.